--- CreativeIdeasBackend/OrdersApi/Logic/Shared/JSONResponse.php
@@ -1,8 +1,0 @@
-<?php
-class JSONResponse {
-    public function render($data) {
-        header('Content-Type: application/json');
-        echo json_encode($data);
-    }
-}
-?>
