+++ OrdersApi/Logic/Shared/JSONResponse.php
@@ -1,0 +1,8 @@
+<?php
+class JSONResponse {
+    public function render($data) {
+        header('Content-Type: application/json');
+        echo json_encode($data);
+    }
+}
+?>
