+++ UsersApi/Logic/EndpointScripts/Test.php
@@ -1,0 +1,7 @@
+<?php
+require __DIR__ . '/../Shared/JSONResponse.php';
+
+$view = new JSONResponse();
+$view->render("This is a test response from the UsersApi!")
+
+?>
